pyproject.toml
[project]
name = "amongllms"
version = "0.1.0"
description = "Among Us as a sandbox for measuring and detecting agentic deception in LLMs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core game engine
    "amongagents",
    # Human trials server
    "fastapi>=0.103.1",
    "uvicorn[standard]>=0.23.2",
    "pydantic>=2.3.0",
    # Utilities
    "python-dotenv>=1.0.1",
    "tqdm>=4.66.2",
]

[project.optional-dependencies]
# For running analysis notebooks in reports/
analysis = [
    "numpy>=2.0.2",
    "pandas>=2.2.2",
    "matplotlib>=3.9.1",
    "seaborn>=0.13.2",
    "openai>=1.32.0",
]
# For running linear probe experiments
probes = [
    "torch>=2.5.1",
    "transformer-lens>=2.11.0",
    "numpy>=2.0.2",
    "pandas>=2.2.2",
]
# Install everything
all = [
    "amongllms[analysis,probes]",
]

[tool.uv.sources]
amongagents = { path = "among-agents" }

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.13",
]

[tool.pytest.ini_options]
# Skip integration tests by default (requires API keys, external services)
# Run integration tests with: pytest -m integration
addopts = "-m 'not integration'"
markers = [
    "integration: marks tests that require external API calls (e.g., OpenRouter)",
]
asyncio_default_fixture_loop_scope = "function"
